// src/types/account.type.ts
export interface IRightMenu {
  myAccount: string;
  editAccount: string;
  password: string;
  addressBook: string;
  wishList: string;
  orderHistory: string;
  downloads: string;
  recurring: string;
  returns: string;
  transactions: string;
  newsletter: string;
  logout: string;
}

export interface ILeftMenu {
  editInfo: string;
  changePwd: string;
  addressBook: string;
  wishList: string;
  orderHistory: string;
  downloads: string;
  rewardPoints: string;
  returns: string;
  transactions: string;
  recurring: string;
  affiliate: string;
  newsletter: string;
}